requirements.txt
django>=5.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-environ
redis
celery[redis]
drf-spectacular
django-extensions
django-redis
stripe
Pillow